Extract `DatagramSendChannelBase`
